requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Werkzeug==2.3.7
python-dotenv==1.0.0
openai==1.3.0
sendgrid==6.10.0
requests==2.31.0
psycopg2-binary>=2.9.9
